Fix some issues with date fields going outside boundaries, split apart filter components to allow better composition
