+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/fields/date-around.tsx
@@ -110,10 +110,8 @@ export const DateAroundField = ({ value, onChange }: DateAroundProps) => {
             ...EnterKeySubmitProps,
           }}
           popoverProps={{
-            modifiers: {
-              preventOverflow: { enabled: false },
-              hide: { enabled: false },
-            },
+            boundary: 'viewport',
+            position: 'bottom',
             onClose: () => {
               setTimeout(() => {
                 blueprintDateRef.current?.setState({ isOpen: false })
