+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/fields/date.tsx
@@ -101,10 +101,8 @@ export const DateField = ({
           ...EnterKeySubmitProps,
         }}
         popoverProps={{
-          modifiers: {
-            preventOverflow: { enabled: false },
-            hide: { enabled: false },
-          },
+          boundary: 'viewport',
+          position: 'bottom',
           onClose: () => {
             setTimeout(() => {
               blueprintDateRef.current?.setState({ isOpen: false })
