+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/filter-builder/filter-leaf.tsx
@@ -4,7 +4,7 @@ import Button from '@mui/material/Button'
 import { useTheme } from '@mui/material/styles'
 import { HoverButton } from '../button/hover'
 import { FilterClass } from './filter.structure'
-import Filter from '../../react-component/filter/filter'
+import FilterPropertyAutocomplete from '../../react-component/filter/filter'
 import { Memo } from '../memo/memo'
 import { ValidationResult } from '../../react-component/location/validators'
 
@@ -16,7 +16,13 @@ type Props = {
   }) => void
 }
 
-const FilterLeaf = ({ filter, setFilter, errorListener }: Props) => {
+export const FilterNegationControls = ({
+  filter,
+  setFilter,
+  children,
+}: Props & {
+  children: React.ReactNode
+}) => {
   const [hover, setHover] = React.useState(false)
   const theme = useTheme()
   return (
@@ -80,15 +86,23 @@ const FilterLeaf = ({ filter, setFilter, errorListener }: Props) => {
           borderColor: theme.palette.primary.main,
         }}
       >
-        <Memo dependencies={[filter, setFilter]}>
-          <Filter
-            filter={filter}
-            setFilter={setFilter}
-            errorListener={errorListener}
-          />
-        </Memo>
+        {children}
       </div>
     </div>
+  )
+}
+
+const FilterLeaf = ({ filter, setFilter, errorListener }: Props) => {
+  return (
+    <FilterNegationControls filter={filter} setFilter={setFilter}>
+      <Memo dependencies={[filter, setFilter]}>
+        <FilterPropertyAutocomplete
+          filter={filter}
+          setFilter={setFilter}
+          errorListener={errorListener}
+        />
+      </Memo>
+    </FilterNegationControls>
   )
 }
 export default hot(module)(FilterLeaf)
